pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-filer"
version = "0.1.0"
description = "Local PDF auto-filing on macOS with Vision OCR + multi-stage Ollama classification + SQLite logging."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="You"}]
dependencies = [
  "pymupdf>=1.24.0",
  "PyYAML>=6.0.1",
  "requests>=2.31.0",
  "rich>=13.7.0",
  "pyobjc-core>=10.0",
  "pyobjc-framework-Vision>=10.0",
  "pyobjc-framework-Quartz>=10.0",
  "pyobjc-framework-Cocoa>=10.0",
]

[project.scripts]
pdf-filer = "pdf_filer.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
